Append Test to the names of the fallbackMethod test classes
