+++ tck/src/main/java/org/eclipse/microprofile/fault/tolerance/tck/fallbackmethod/FallbackMethodGenericArrayTest.java
@@ -24,8 +24,8 @@ import static org.testng.Assert.assertEquals;
 
 import javax.inject.Inject;
 
-import org.eclipse.microprofile.fault.tolerance.tck.fallbackmethod.beans.FallbackMethodWithinInterfaceA;
-import org.eclipse.microprofile.fault.tolerance.tck.fallbackmethod.beans.FallbackMethodWithinInterfaceB;
+import org.eclipse.microprofile.fault.tolerance.tck.fallbackmethod.beans.FallbackMethodGenericArrayBeanA;
+import org.eclipse.microprofile.fault.tolerance.tck.fallbackmethod.beans.FallbackMethodGenericArrayBeanB;
 import org.jboss.arquillian.container.test.api.Deployment;
 import org.jboss.arquillian.testng.Arquillian;
 import org.jboss.shrinkwrap.api.ShrinkWrap;
@@ -35,28 +35,27 @@ import org.jboss.shrinkwrap.api.spec.WebArchive;
 import org.testng.annotations.Test;
 
 /**
- * Test for a fallback method defined on an interface and implemented in a subclass
+ * Test for a fallback method with a generic array parameter
  */
-public class FallbackMethodWithinInterface extends Arquillian {
-    
+public class FallbackMethodGenericArrayTest extends Arquillian {
+
     @Deployment
     public static WebArchive deploy() {
-        JavaArchive testJar = ShrinkWrap.create(JavaArchive.class, "ftFallbackMethodWithinInterface.jar")
-                .addClasses(FallbackMethodWithinInterfaceA.class,
-                            FallbackMethodWithinInterfaceB.class)
+        JavaArchive testJar = ShrinkWrap.create(JavaArchive.class, "ftFallbackMethodGenericArray.jar")
+                .addClasses(FallbackMethodGenericArrayBeanA.class, FallbackMethodGenericArrayBeanB.class)
                 .addAsManifestResource(EmptyAsset.INSTANCE, "beans.xml");
         
         WebArchive war = ShrinkWrap
-                .create(WebArchive.class, "ftFallbackMethodWithinInterface.war")
+                .create(WebArchive.class, "ftFallbackMethodGenericArray.war")
                 .addAsLibrary(testJar);
         return war;
     }
     
-    @Inject private FallbackMethodWithinInterfaceA bean;
+    @Inject private FallbackMethodGenericArrayBeanA bean;
     
     @Test
-    public void fallbackMethodWithinInterface() {
-        assertEquals(bean.method(1, 2L), "fallback");
+    public void fallbackMethodGenericArray() {
+        assertEquals(bean.method(new String[3][4]), "fallback");
     }
-
+    
 }
