+++ tck/src/main/java/org/eclipse/microprofile/fault/tolerance/tck/fallbackmethod/FallbackMethodGenericTest.java
@@ -24,8 +24,8 @@ import static org.testng.Assert.assertEquals;
 
 import javax.inject.Inject;
 
-import org.eclipse.microprofile.fault.tolerance.tck.fallbackmethod.beans.FallbackMethodDefaultMethodA;
-import org.eclipse.microprofile.fault.tolerance.tck.fallbackmethod.beans.FallbackMethodDefaultMethodB;
+import org.eclipse.microprofile.fault.tolerance.tck.fallbackmethod.beans.FallbackMethodGenericBeanA;
+import org.eclipse.microprofile.fault.tolerance.tck.fallbackmethod.beans.FallbackMethodGenericBeanB;
 import org.jboss.arquillian.container.test.api.Deployment;
 import org.jboss.arquillian.testng.Arquillian;
 import org.jboss.shrinkwrap.api.ShrinkWrap;
@@ -35,28 +35,27 @@ import org.jboss.shrinkwrap.api.spec.WebArchive;
 import org.testng.annotations.Test;
 
 /**
- * Test for a fallback method defined as a default method on an interface
+ * Test for a fallback methods with a type variable parameter
  */
-public class FallbackMethodDefaultMethod extends Arquillian {
-    
+public class FallbackMethodGenericTest extends Arquillian {
+
     @Deployment
     public static WebArchive deploy() {
-        JavaArchive testJar = ShrinkWrap.create(JavaArchive.class, "ftFallbackMethodDefaultMethod.jar")
-                .addClasses(FallbackMethodDefaultMethodA.class,
-                            FallbackMethodDefaultMethodB.class)
+        JavaArchive testJar = ShrinkWrap.create(JavaArchive.class, "ftFallbackMethodGeneric.jar")
+                .addClasses(FallbackMethodGenericBeanA.class, FallbackMethodGenericBeanB.class)
                 .addAsManifestResource(EmptyAsset.INSTANCE, "beans.xml");
         
         WebArchive war = ShrinkWrap
-                .create(WebArchive.class, "ftFallbackMethodDefaultMethod.war")
+                .create(WebArchive.class, "ftFallbackMethodGeneric.war")
                 .addAsLibrary(testJar);
         return war;
     }
     
-    @Inject private FallbackMethodDefaultMethodA bean;
+    @Inject private FallbackMethodGenericBeanA bean;
     
     @Test
-    public void fallbackMethodDefaultMethod() {
+    public void fallbackMethodGeneric() {
         assertEquals(bean.method(1, 2L), "fallback");
     }
-
+    
 }
