+++ tck/src/main/java/org/eclipse/microprofile/fault/tolerance/tck/fallbackmethod/FallbackMethodSuperclassTest.java
@@ -24,8 +24,8 @@ import static org.testng.Assert.assertEquals;
 
 import javax.inject.Inject;
 
-import org.eclipse.microprofile.fault.tolerance.tck.fallbackmethod.beans.FallbackMethodGenericBeanA;
-import org.eclipse.microprofile.fault.tolerance.tck.fallbackmethod.beans.FallbackMethodGenericBeanB;
+import org.eclipse.microprofile.fault.tolerance.tck.fallbackmethod.beans.FallbackMethodSuperclassBeanA;
+import org.eclipse.microprofile.fault.tolerance.tck.fallbackmethod.beans.FallbackMethodSuperclassBeanB;
 import org.jboss.arquillian.container.test.api.Deployment;
 import org.jboss.arquillian.testng.Arquillian;
 import org.jboss.shrinkwrap.api.ShrinkWrap;
@@ -35,27 +35,27 @@ import org.jboss.shrinkwrap.api.spec.WebArchive;
 import org.testng.annotations.Test;
 
 /**
- * Test for a fallback methods with a type variable parameter
+ * Test for protected fallback method on a superclass
  */
-public class FallbackMethodGeneric extends Arquillian {
-
+public class FallbackMethodSuperclassTest extends Arquillian {
+    
     @Deployment
     public static WebArchive deploy() {
-        JavaArchive testJar = ShrinkWrap.create(JavaArchive.class, "ftFallbackMethodGeneric.jar")
-                .addClasses(FallbackMethodGenericBeanA.class, FallbackMethodGenericBeanB.class)
+        JavaArchive testJar = ShrinkWrap.create(JavaArchive.class, "ftFallbackMethodSuperclass.jar")
+                .addClasses(FallbackMethodSuperclassBeanA.class, FallbackMethodSuperclassBeanB.class)
                 .addAsManifestResource(EmptyAsset.INSTANCE, "beans.xml");
         
         WebArchive war = ShrinkWrap
-                .create(WebArchive.class, "ftFallbackMethodGeneric.war")
+                .create(WebArchive.class, "ftFallbackMethodSuperclass.war")
                 .addAsLibrary(testJar);
         return war;
     }
     
-    @Inject private FallbackMethodGenericBeanA bean;
+    @Inject private FallbackMethodSuperclassBeanA bean;
     
     @Test
-    public void fallbackMethodGeneric() {
+    public void fallbackMethodSuperclass() {
         assertEquals(bean.method(1, 2L), "fallback");
     }
-    
+
 }
